Bug corrections, UI improved and optimization

+++ app/src/main/java/com/nico_11_riv/intranetepitech/ProfileActivityFragment.java
@@ -9,6 +9,7 @@ import android.util.Log;
 import android.view.View;
 import android.widget.ImageView;
 import android.widget.ProgressBar;
+import android.widget.RelativeLayout;
 import android.widget.TextView;
 import android.widget.Toast;
 
@@ -16,11 +17,12 @@ import com.nico_11_riv.intranetepitech.api.APIErrorHandler;
 import com.nico_11_riv.intranetepitech.api.IntrAPI;
 import com.nico_11_riv.intranetepitech.database.Message;
 import com.nico_11_riv.intranetepitech.database.Userinfos;
-import com.nico_11_riv.intranetepitech.toolbox.CircleTransform;
-import com.nico_11_riv.intranetepitech.database.setters.user.GUserInfos;
-import com.nico_11_riv.intranetepitech.database.setters.user.PUserInfos;
 import com.nico_11_riv.intranetepitech.database.setters.messages.PMessages;
 import com.nico_11_riv.intranetepitech.database.setters.user.GUser;
+import com.nico_11_riv.intranetepitech.database.setters.user.GUserInfos;
+import com.nico_11_riv.intranetepitech.database.setters.user.PUserInfos;
+import com.nico_11_riv.intranetepitech.toolbox.CircleTransform;
+import com.nico_11_riv.intranetepitech.toolbox.IsConnected;
 import com.nico_11_riv.intranetepitech.ui.adapters.RVMessagesAdapter;
 import com.nico_11_riv.intranetepitech.ui.contents.MessageContent;
 import com.orm.query.Condition;
@@ -79,9 +81,13 @@ public class ProfileActivityFragment extends Fragment {
     @ViewById
     ProgressBar messages_progress;
 
+    @ViewById
+    RelativeLayout rl;
+
     private GUser gUser = new GUser();
     private GUserInfos user_info;
     private RVMessagesAdapter adapter;
+    private IsConnected ic;
 
     @AfterInject
     void afterInject() {
@@ -97,14 +103,14 @@ public class ProfileActivityFragment extends Fragment {
 
     @UiThread
     void filluserinfosui() {
-        Picasso.with(getContext()).load(user_info.getPicture()).transform(new CircleTransform()).into(student_img);
+        Picasso.with(getActivity().getApplicationContext()).load(user_info.getPicture()).transform(new CircleTransform()).into(student_img);
         login.setText(user_info.getLogin());
         TextView tv = (TextView)getActivity().findViewById(R.id.menu_login);
         tv.setText(user_info.getLogin());
         tv = (TextView)getActivity().findViewById(R.id.menu_email);
         tv.setText(user_info.getEmail());
         ImageView iv = (ImageView)getActivity().findViewById(R.id.menu_img);
-        Picasso.with(getContext()).load(user_info.getPicture()).transform(new CircleTransform()).into(iv);
+        Picasso.with(getActivity().getApplicationContext()).load(user_info.getPicture()).transform(new CircleTransform()).into(iv);
         email.setText(user_info.getEmail());
         gpa_content.setText(user_info.getGpa());
         promo_content.setText(user_info.getPromo());
@@ -125,58 +131,70 @@ public class ProfileActivityFragment extends Fragment {
 
     @UiThread
     void fillnewmessagesui() {
-        messages_progress.setVisibility(View.GONE);
         adapter.background();
     }
 
     void setUserInfos() {
         List <Userinfos> uInfos = Userinfos.findWithQuery(Userinfos.class, "SELECT * FROM Userinfos WHERE login = ?", gUser.getLogin());
+
         if (uInfos.size() > 0) {
             user_info = new GUserInfos();
             filluserinfosui();
         }
-       Userinfos.deleteAll(Userinfos.class, "login = ?", gUser.getLogin());
-        api.setCookie("PHPSESSID", gUser.getToken());
-        try {
-            PUserInfos infos = new PUserInfos();
-            infos.init(api.getuserinfo(gUser.getLogin()));
-        } catch (HttpClientErrorException e) {
-            Log.d("Response", e.getResponseBodyAsString());
-            Toast.makeText(getContext(), "Erreur de l'API", Toast.LENGTH_SHORT).show();
-        } catch (NullPointerException e) {
-            Toast.makeText(getContext(), "Erreur de l'API", Toast.LENGTH_SHORT).show();
-            e.printStackTrace();
+        if (ic.connected()) {
+            Userinfos.deleteAll(Userinfos.class, "login = ?", gUser.getLogin());
+            api.setCookie("PHPSESSID", gUser.getToken());
+            try {
+                PUserInfos infos = new PUserInfos();
+                infos.init(api.getuserinfo(gUser.getLogin()));
+            } catch (HttpClientErrorException e) {
+                Log.d("Response", e.getResponseBodyAsString());
+                Toast.makeText(getActivity().getApplicationContext(), e.getMessage(), Toast.LENGTH_SHORT).show();
+            } catch (NullPointerException e) {
+                Toast.makeText(getActivity().getApplicationContext(), e.getMessage(), Toast.LENGTH_SHORT).show();
+                e.printStackTrace();
+            }
+            user_info = new GUserInfos();
+            filluserinfosui();
         }
-        user_info = new GUserInfos();
-        filluserinfosui();
     }
 
     void setMessages() {
         fillmessagesui();
-        api.setCookie("PHPSESSID", gUser.getToken());
-        try {
-            PMessages msgs = new PMessages();
-            msgs.init(api.getnotifs());
-        } catch (HttpClientErrorException e) {
-            Log.d("Response", e.getResponseBodyAsString());
-            Toast.makeText(getContext(), "Erreur de l'API", Toast.LENGTH_SHORT).show();
-        } catch (NullPointerException e) {
-            Toast.makeText(getContext(), "Erreur de l'API", Toast.LENGTH_SHORT).show();
-            e.printStackTrace();
+        if (ic.connected()) {
+            api.setCookie("PHPSESSID", gUser.getToken());
+            try {
+                PMessages msgs = new PMessages();
+                msgs.init(api.getnotifs());
+            } catch (HttpClientErrorException e) {
+                Log.d("Response", e.getResponseBodyAsString());
+                Toast.makeText(getActivity().getApplicationContext(), e.getMessage(), Toast.LENGTH_SHORT).show();
+            } catch (NullPointerException e) {
+                Toast.makeText(getActivity().getApplicationContext(), e.getMessage(), Toast.LENGTH_SHORT).show();
+                e.printStackTrace();
+            }
+            fillnewmessagesui();
         }
-        fillnewmessagesui();
+    }
+
+    @UiThread
+    void setProgressBar() {
+        messages_progress.setVisibility(View.GONE);
     }
 
     @Background
     void profile_messages() {
         setUserInfos();
         setMessages();
+        setProgressBar();
     }
 
     @AfterViews
     void init() {
-        if (getResources().getConfiguration().orientation == Configuration.ORIENTATION_LANDSCAPE)
+        ic = new IsConnected(getActivity().getApplicationContext());
+        if (getResources().getConfiguration().orientation == Configuration.ORIENTATION_LANDSCAPE) {
             background.setVisibility(View.GONE);
+        }
         LinearLayoutManager llm = new LinearLayoutManager(getActivity());
         rv.setLayoutManager(llm);
         profile_messages();
